shell: tunnel docker connection during client dial
Instead of manually setting up an SSH tunnel and forwarding docker
socket traffic, just override the default dialer in the Docker client
with one that sets up the tunnel for us. This eliminates the race
condition that was affecting the previous proxy route.

Inspired by https://the.binbashtheory.com/docker-tunneling-in-go/, but
needed various modifications to get this working with
fsouza/go-dockerclient. I also tried with the upstream docker client
library, but that's not compatible with the dockerpty library. However I
ended up having to fork the dockerpty library anyway to fix a stdin
issue there..